pyproject.toml
[project]
name = "ai-agent-rag"
version = "0.1.0"
description = ""
authors = [{ name = "Alexsander Valente", email = "alevtelles@gmail.com" }]
requires-python = ">=3.13,<4.0.0"
dependencies = ["langchain (>=1.2.10,<2.0.0)", "langchain-community (>=0.4.1,<0.5.0)", "langchain-text-splitters (>=1.1.0,<2.0.0)", "langchain-openai (>=1.1.9,<2.0.0)", "langchain-chroma (>=1.1.0,<2.0.0)", "python-dotenv (>=1.2.1,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
